refactor: Moved Contact type to own file
