+++ src/utils/types/class.ts
@@ -1,3 +1,4 @@
+import { Contact } from "./contact";
 import { Student, Teacher } from "./person";
 
 export type Class = {
@@ -9,19 +10,4 @@ export type Class = {
   classAdvisors: Array<Teacher>;
   contacts: Array<Contact>;
   students: Array<Student>;
-};
-
-export type Contact = {
-  id: number;
-  name: {
-    "en-US": string;
-    th: string;
-  };
-  via: "facebook" | "line" | "discord" | "email";
-  includes: {
-    students: boolean;
-    parents: boolean;
-    teachers: boolean;
-  };
-  url: string;
 };
